Resolve ArchUnit exemptions for CountryController and EmployeeController
Extracted repository access out of CountryController/EmployeeController into
new internal CountryService/EmployeeService classes, following the existing
service-layer pattern used elsewhere in the codebase. This removes their
need for an exemption from the "rest controllers should not expose database
entities directly" rule, and lets the previously-proposed "rest controllers
should not depend on database entities directly" rule be added without
exempting them.
