+++ backend/src/main/kotlin/at/wrk/tafel/admin/backend/modules/base/country/CountryController.kt
@@ -1,6 +1,6 @@
 package at.wrk.tafel.admin.backend.modules.base.country
 
-import at.wrk.tafel.admin.backend.database.model.staticdata.CountryRepository
+import at.wrk.tafel.admin.backend.modules.base.country.internal.CountryService
 import org.springframework.security.access.prepost.PreAuthorize
 import org.springframework.web.bind.annotation.GetMapping
 import org.springframework.web.bind.annotation.RequestMapping
@@ -10,17 +10,11 @@ import org.springframework.web.bind.annotation.RestController
 @RequestMapping("/api/countries")
 @PreAuthorize("isAuthenticated()")
 class CountryController(
-    private val countryRepository: CountryRepository,
+    private val countryService: CountryService,
 ) {
 
     @GetMapping
     fun listCountries(): CountryListResponse = CountryListResponse(
-        items = countryRepository.findAll().map {
-            Country(
-                id = it.id!!,
-                code = it.code!!,
-                name = it.name!!,
-            )
-        },
+        items = countryService.listCountries(),
     )
 }
