+++ backend/src/test/kotlin/at/wrk/tafel/admin/backend/architecture/ProjectSpecificRulesTest.kt
@@ -1,12 +1,14 @@
 package at.wrk.tafel.admin.backend.architecture
 
 import at.wrk.tafel.admin.backend.architecture.conditions.HaveApiRequestMappingPathCondition
-import at.wrk.tafel.admin.backend.modules.base.country.CountryController
-import at.wrk.tafel.admin.backend.modules.base.employee.EmployeeController
+import at.wrk.tafel.admin.backend.modules.distribution.DistributionController
+import at.wrk.tafel.admin.backend.modules.distribution.internal.ticket.DistributionTicketController
+import at.wrk.tafel.admin.backend.modules.distribution.internal.ticket.DistributionTicketScreenController
 import com.tngtech.archunit.core.domain.JavaClass.Predicates.resideInAPackage
 import com.tngtech.archunit.junit.AnalyzeClasses
 import com.tngtech.archunit.junit.ArchTest
 import com.tngtech.archunit.lang.syntax.ArchRuleDefinition.classes
+import com.tngtech.archunit.lang.syntax.ArchRuleDefinition.noClasses
 import com.tngtech.archunit.lang.syntax.ArchRuleDefinition.noMethods
 import org.springframework.web.bind.annotation.RestController
 
@@ -22,11 +24,19 @@ internal class ProjectSpecificRulesTest {
     val `rest controllers should not expose database entities directly` = noMethods()
         .that().arePublic()
         .and().areDeclaredInClassesThat().areAnnotatedWith(RestController::class.java)
-        .and().areDeclaredInClassesThat().areNotAssignableTo(CountryController::class.java)
-        .and().areDeclaredInClassesThat().areNotAssignableTo(EmployeeController::class.java)
         .should().haveRawReturnType(resideInAPackage("..database.model.."))
+        .because("controllers should map entities to DTOs instead of returning them directly from an endpoint")
+
+    @ArchTest
+    val `rest controllers should not depend on database entities directly` = noClasses()
+        .that().areAnnotatedWith(RestController::class.java)
+        .and().areNotAssignableTo(DistributionController::class.java)
+        .and().areNotAssignableTo(DistributionTicketController::class.java)
+        .and().areNotAssignableTo(DistributionTicketScreenController::class.java)
+        .should().dependOnClassesThat().resideInAPackage("..database.model..")
         .because(
-            "controllers should map entities to DTOs instead of returning them directly from an endpoint - " +
-                "CountryController and EmployeeController are exempted for now as they still access repositories directly",
+            "controllers should map entities to DTOs instead of exposing/depending on them directly - " +
+                "DistributionController, DistributionTicketController and DistributionTicketScreenController " +
+                "are exempted for now as they still access entities directly",
         )
 }
