+++ backend/src/test/kotlin/at/wrk/tafel/admin/backend/modules/base/country/CountryControllerTest.kt
@@ -1,7 +1,6 @@
 package at.wrk.tafel.admin.backend.modules.base.country
 
-import at.wrk.tafel.admin.backend.database.model.staticdata.CountryEntity
-import at.wrk.tafel.admin.backend.database.model.staticdata.CountryRepository
+import at.wrk.tafel.admin.backend.modules.base.country.internal.CountryService
 import io.mockk.every
 import io.mockk.impl.annotations.InjectMockKs
 import io.mockk.impl.annotations.RelaxedMockK
@@ -14,35 +13,22 @@ import org.junit.jupiter.api.extension.ExtendWith
 class CountryControllerTest {
 
     @RelaxedMockK
-    private lateinit var countryRepository: CountryRepository
+    private lateinit var countryService: CountryService
 
     @InjectMockKs
     private lateinit var countryController: CountryController
 
     @Test
     fun `list countries`() {
-        val country1 = CountryEntity()
-        country1.id = 1
-        country1.code = "AA"
-        country1.name = "Name A"
+        val country1 = Country(id = 1, code = "AA", name = "Name A")
+        val country2 = Country(id = 2, code = "BB", name = "Name B")
 
-        val country2 = CountryEntity()
-        country2.id = 2
-        country2.code = "BB"
-        country2.name = "Name B"
-
-        every { countryRepository.findAll() } returns listOf(country1, country2)
+        every { countryService.listCountries() } returns listOf(country1, country2)
 
         val response = countryController.listCountries()
 
-        assertThat(response).isNotNull
-        assertThat(response.items).hasSize(2)
-
-        assertThat(response.items).hasSameElementsAs(
-            listOf(
-                Country(id = 1, code = "AA", name = "Name A"),
-                Country(id = 2, code = "BB", name = "Name B"),
-            ),
+        assertThat(response).isEqualTo(
+            CountryListResponse(items = listOf(country1, country2)),
         )
     }
 }
